*: move pkg/file to pkg/utils
